Cleaned up utility class, ExtractMinimalConnectedNetworkFromOSMUtilities
